Modularize events type definition
Also:
* Fix extra whitespace in rhino/file/module.fifty.ts
