+++ loader/$api.fifty.ts
@@ -99,32 +99,6 @@ namespace slime.$api {
 		threads: any
 	}
 
-	/**
-	 * An occurrence in which other parts of a program might be interested. Events have a
-	 */
-	export interface Event<T> {
-		type: string
-		source: object
-		timestamp: number
-		detail: T
-		path: any[]
-	}
-
-	export namespace event {
-		/**
-		 * A function that receives events.
-		 */
-		export type Handler<T> = (e: Event<T>) => void
-	}
-
-	export interface Events<D> {
-		listeners: {
-			add: <K extends keyof D>(type: K, handler: event.Handler<D[K]>) => void
-			remove: <K extends keyof D>(type: K, handler: event.Handler<D[K]>) => void
-		},
-		fire: <K extends keyof D>(type: K, detail?: D[K]) => void
-	}
-
 	export namespace events {
 		/**
 		 * An object whose methods process events; events of a type are mapped to a method with the same name as that type.
